Polymorphic constraint solving is not implemented yet - no error returned in t3
This refers to section _5.1.1 Greedy instantiation_, specifically to the quote:

> This works perfectly well when the second argument has the same type,
> or when the second argument is of a subtype of Cat (e.g. Tabby),
> but fails when the second argument is of a larger type.

Function t3:p/2 in test/0 is called with the second argument being a supertype of the first.
I was curious to see whether this leads to an instantiation error in Gradualizer.
Apparently, no solving for the actual type is present in Gradualizer yet.